Add: enable rider self-removal from campaign
If a rider signs up for any tasks, we need to ensure that if they remove
themselves, it also removes them from the RiderCampaign table.

Most of this commit is navigating the handle_event callbacks for the
events "Signup" and "remove me as rider".

I updated `Delivery.remove_rider_from_campaign` to take a rider `id`
instead of the entire rider so it could be better shared across
files.