Add unfilled tasks / total tasks in campaign index
This displays the number of open tasks against the total number of
tasks, for example: "3 / 5 Tasks filled".

I thought I could use the campaign_stats view, but I don't think there
is a column to show "unassigned tasks"... and I don't think it's safe to
assume that `# tasks - # of riders` will give us an accurate picture of
things since it's possible for a rider to take on multiple tasks.

So, I added a preload of :tasks to campaigns to be able to manually
filter that information.

We could maybe alter the campaign_stats for this data rather than do a
preload, but I didn't want to go deep on that as it would take longer
that this.