Refactor: remove_rider_from_campaign parameters.
We will need to use this function from the rider signup view for riders
to be able to unassign themselves.

The "current_user" only has a "rider_id", which would require us to
preload the rider or fetch the current_user's rider in
order to use this function in its current state.

Since the query only makes use of the rider's id, we can just pass that.