.github,etc: enforce `mix format` in CI
Consistent and automated code formatting can make it much easier to
work with a codebase.

In #396 a run of `mix format` with args was done.

This change enforces use of `mix format` in CI and adds a note to the
README about how to format. A hint for editor and git config is also
added.

The existing formatter config is fixed up so just running `mix format`
should do the right thing.

Updates #394